+++ src/pyramid/scripts/common.py
@@ -1,4 +1,6 @@
 import plaster
+import plaster.exceptions
+import sys
 
 
 def parse_vars(args):
@@ -15,9 +17,23 @@ def parse_vars(args):
     return result
 
 
-def get_config_loader(config_uri):
-    """
-    Find a ``plaster.ILoader`` object supporting the "wsgi" protocol.
+def get_config_loader(config_uri, report_failure=None):
+    """Find a ``plaster.ILoader`` object supporting the "wsgi" protocol.
+
+    ``report_failure``, if passed, must be a function of one argument,
+    a string.  The function is called with an error message when the
+    ``config_uri`` cannot be used to obtain settings.  After
+    ``report_failure`` is called the program exits with a ``1``
+    (failure) status code.
+
+    When ``report_failure`` is not passed, or is None, the caller is expected
+    to handle PlasterError exceptions raised by :term:`plaster`.
 
     """
-    return plaster.get_loader(config_uri, protocols=['wsgi'])
+    try:
+        return plaster.get_loader(config_uri, protocols=['wsgi'])
+    except plaster.exceptions.PlasterError as e:
+        if not report_failure:
+            raise e
+        report_failure(f'The settings given are not available: {e}')
+        sys.exit(1)